requirements.txt
gtts==2.2.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-api-python-client==2.33.0
moviepy==1.0.3
schedule==1.1.0
pytrends==4.8.0